# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: WhelpWise Breeder Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stud-listings:
    get:
      operationId: listStudListings
      tags: [stud-listings]
      summary: List public stud listings
      parameters:
        - name: breed
          in: query
          schema:
            type: string
        - name: country
          in: query
          schema:
            type: string
        - name: healthStatus
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Stud listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudListing"
    post:
      operationId: createStudListing
      tags: [stud-listings]
      summary: Create a stud listing for a dog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudListingInput"
      responses:
        "201":
          description: Stud listing created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudListing"
  /stud-listings/{listingId}:
    get:
      operationId: getStudListing
      tags: [stud-listings]
      summary: Get a stud listing
      parameters:
        - name: listingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Stud listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudListing"
    patch:
      operationId: updateStudListing
      tags: [stud-listings]
      summary: Update a stud listing
      parameters:
        - name: listingId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudListingUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudListing"
    delete:
      operationId: deleteStudListing
      tags: [stud-listings]
      summary: Delete a stud listing
      parameters:
        - name: listingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Waiting List ────────────────────────────────────────────────────────────
...
  /pets:
    get:
      operationId: listPets
      tags: [pets]
      summary: List family pets
      responses:
        "200":
          description: Pet list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FamilyPet"
    post:
      operationId: createPet
      tags: [pets]
      summary: Add a family pet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FamilyPetInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FamilyPet"
  /pets/{petId}:
    get:
      operationId: getPet
      tags: [pets]
      summary: Get a family pet
      parameters:
        - name: petId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Pet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FamilyPet"
    put:
      operationId: updatePet
      tags: [pets]
      summary: Update a family pet
      parameters:
        - name: petId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FamilyPetUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FamilyPet"
    delete:
      operationId: deletePet
      tags: [pets]
      summary: Delete a family pet
      parameters:
        - name: petId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /pets/{petId}/vaccinations:
    get:
      operationId: listPetVaccinations
      tags: [pets]
      summary: List vaccinations for a pet
      parameters:
        - name: petId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vaccination list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PetVaccination"
    post:
      operationId: createPetVaccination
      tags: [pets]
      summary: Log a vaccination
      parameters:
        - name: petId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PetVaccinationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PetVaccination"
  /pets/{petId}/vaccinations/{vaccinationId}:
    delete:
      operationId: deletePetVaccination
      tags: [pets]
      summary: Delete a vaccination record
      parameters:
        - name: petId
          in: path
          required: true
          schema:
            type: integer
        - name: vaccinationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /pets/{petId}/vet-visits:
    get:
      operationId: listPetVetVisits
      tags: [pets]
      summary: List vet visits for a pet
      parameters:
        - name: petId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vet visit list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PetVetVisit"
    post:
      operationId: createPetVetVisit
      tags: [pets]
      summary: Log a vet visit
      parameters:
        - name: petId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PetVetVisitInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PetVetVisit"
  /pets/{petId}/vet-visits/{visitId}:
    delete:
      operationId: deletePetVetVisit
      tags: [pets]
      summary: Delete a vet visit
      parameters:
        - name: petId
          in: path
          required: true
          schema:
            type: integer
        - name: visitId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    StudListing:
      type: object
      properties:
        id:
          type: integer
        dogId:
          type: integer
        dogName:
          type: string
        breedName:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        studFee:
          type: ["number", "null"]
        currency:
          type: string
        country:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        healthTested:
          type: boolean
        active:
          type: boolean
        expiresAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - dogId
        - dogName
        - currency
        - healthTested
        - active
        - createdAt
    StudListingInput:
      type: object
      properties:
        dogId:
          type: integer
        studFee:
          type: ["number", "null"]
        currency:
          type: string
        country:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        expiresAt:
          type: ["string", "null"]
      required:
        - dogId
        - currency
    StudListingUpdate:
      type: object
      properties:
        studFee:
          type: ["number", "null"]
        currency:
          type: string
        country:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        active:
          type: boolean
        expiresAt:
          type: ["string", "null"]
...
    FamilyPet:
      type: object
      required: [id, name, species, status]
      properties:
        id:
          type: integer
        name:
          type: string
        species:
          type: string
        breed:
          type: ["string", "null"]
        sex:
          type: ["string", "null"]
        dob:
          type: ["string", "null"]
        colour:
          type: ["string", "null"]
        microchip:
          type: ["string", "null"]
        vetName:
          type: ["string", "null"]
        vetPhone:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string
    FamilyPetInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        species:
          type: string
        breed:
          type: ["string", "null"]
        sex:
          type: ["string", "null"]
        dob:
          type: ["string", "null"]
        colour:
          type: ["string", "null"]
        microchip:
          type: ["string", "null"]
        vetName:
          type: ["string", "null"]
        vetPhone:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        status:
          type: string
    FamilyPetUpdate:
      type: object
      properties:
        name:
          type: string
        species:
          type: string
        breed:
          type: ["string", "null"]
        sex:
          type: ["string", "null"]
        dob:
          type: ["string", "null"]
        colour:
          type: ["string", "null"]
        microchip:
          type: ["string", "null"]
        vetName:
          type: ["string", "null"]
        vetPhone:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        status:
          type: string
    PetVaccination:
      type: object
      required: [id, petId, vaccine]
      properties:
        id:
          type: integer
        petId:
          type: integer
        vaccine:
          type: string
        dateGiven:
          type: ["string", "null"]
        nextDueDate:
          type: ["string", "null"]
        vet:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    PetVaccinationInput:
      type: object
      required: [vaccine]
      properties:
        vaccine:
          type: string
        dateGiven:
          type: ["string", "null"]
        nextDueDate:
          type: ["string", "null"]
        vet:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    PetVetVisit:
      type: object
      required: [id, petId, date, reason]
      properties:
        id:
          type: integer
        petId:
          type: integer
        date:
          type: string
        reason:
          type: string
        vet:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        cost:
          type: ["string", "null"]
        createdAt:
          type: string
    PetVetVisitInput:
      type: object
      required: [date, reason]
      properties:
        date:
          type: string
        reason:
          type: string
        vet:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        cost:
          type: ["string", "null"]